Vault Reference Updates
I propose making two changes:Updating the HashiCorp references to use the correct format as well as specifically calling out support for HCP Vault. This is because on the HashiCorp site for Codefresh (https://www.hashicorp.com/partners/tech/codefresh) we will be creating a new  integration listing for HCP Vault. The getting started button will point to this page and it would be great to specifically mention HCP Vault support to provide clarify for our common users and customers.